refactor: retarget file channel owner
