+++ backend/file_channel.py
@@ -63,11 +63,6 @@ def get_file_channel_binding(thread_id: str) -> FileChannelBinding:
     )
 
 
-# ---------------------------------------------------------------------------
-# File CRUD -- delegates to the local file-channel source
-# ---------------------------------------------------------------------------
-
-
 def save_file(*, thread_id: str, relative_path: str, content: bytes) -> dict:
     """Save file to the thread's file channel."""
     source = get_file_channel_source(thread_id)
