+++ tests/Unit/backend/web/services/test_file_channel_service.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from pathlib import Path
 from types import SimpleNamespace
 
-from backend.web.services import file_channel_service
+from backend import file_channel as file_channel_service
 
 
 class _ThreadRepo:
